+++ src/App/Services.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+use App\Service\Note;
 use App\Service\Note\NoteService;
 use App\Service\Task\TaskService;
 use App\Service\User\UserService;
@@ -29,6 +30,24 @@ $container['note_service'] = static function (
     ContainerInterface $container
 ): NoteService {
     return new NoteService(
+        $container->get('note_repository'),
+        $container->get('redis_service')
+    );
+};
+
+$container['create_note_service'] = static function (
+    ContainerInterface $container
+): Note\Create {
+    return new Note\Create(
+        $container->get('note_repository'),
+        $container->get('redis_service')
+    );
+};
+
+$container['update_note_service'] = static function (
+    ContainerInterface $container
+): Note\Update {
+    return new Note\Update(
         $container->get('note_repository'),
         $container->get('redis_service')
     );
