+++ src/Controller/Note/Create.php
@@ -12,7 +12,7 @@ final class Create extends Base
     public function __invoke(Request $request, Response $response): Response
     {
         $input = (array) $request->getParsedBody();
-        $note = $this->getNoteService()->create($input);
+        $note = $this->getServiceCreateNote()->create($input);
 
         return $this->jsonResponse($response, 'success', $note, 201);
     }
