+++ src/Controller/Note/Update.php
@@ -12,7 +12,7 @@ final class Update extends Base
     public function __invoke(Request $request, Response $response, array $args): Response
     {
         $input = (array) $request->getParsedBody();
-        $note = $this->getNoteService()->update($input, (int) $args['id']);
+        $note = $this->getServiceUpdateNote()->update($input, (int) $args['id']);
 
         return $this->jsonResponse($response, 'success', $note, 200);
     }
