+++ src/Service/Note/NoteService.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace App\Service\Note;
 
-use App\Exception\Note;
-
 final class NoteService extends Base
 {
     public function getAll(): array
@@ -27,43 +25,6 @@ final class NoteService extends Base
     public function search(string $notesName): array
     {
         return $this->noteRepository->searchNotes($notesName);
-    }
-
-    public function create(array $input): object
-    {
-        $data = json_decode(json_encode($input), false);
-        if (! isset($data->name)) {
-            throw new Note('Invalid data: name is required.', 400);
-        }
-        self::validateNoteName($data->name);
-        $data->description = $data->description ?? null;
-        $note = $this->noteRepository->createNote($data);
-        if (self::isRedisEnabled() === true) {
-            $this->saveInCache($note->id, $note);
-        }
-
-        return $note;
-    }
-
-    public function update(array $input, int $noteId): object
-    {
-        $note = $this->getOneFromDb($noteId);
-        $data = json_decode(json_encode($input), false);
-        if (! isset($data->name) && ! isset($data->description)) {
-            throw new Note('Enter the data to update the note.', 400);
-        }
-        if (isset($data->name)) {
-            $note->name = self::validateNoteName($data->name);
-        }
-        if (isset($data->description)) {
-            $note->description = $data->description;
-        }
-        $notes = $this->noteRepository->updateNote($note);
-        if (self::isRedisEnabled() === true) {
-            $this->saveInCache($notes->id, $notes);
-        }
-
-        return $notes;
     }
 
     public function delete(int $noteId): void
